Changes based on feedback from Dave.
