+++ src/main/java/gov/usgs/wma/mlrauthserver/service/SAMLUserDetailsImpl.java
@@ -1,9 +1,14 @@
 package gov.usgs.wma.mlrauthserver.service;
 
 import gov.usgs.wma.mlrauthserver.model.WaterAuthUser;
+import gov.usgs.wma.mlrauthserver.model.WaterAuthUserDetails;
 import gov.usgs.wma.mlrauthserver.util.SAMLUtils;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.List;
+import java.util.Map;
+
+import org.hibernate.validator.constraints.NotEmpty;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.saml.SAMLCredential;
 import org.springframework.security.saml.userdetails.SAMLUserDetailsService;
@@ -11,12 +16,39 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class SAMLUserDetailsImpl implements SAMLUserDetailsService  {
+	@NotEmpty
+	@Value("${security.saml.attribute-names.username}")
+	private String[] samlUsernameAttributeNames;
 
-	@Autowired
-	WaterAuthUserFactory waterAuthUserFactory;
+	@NotEmpty
+	@Value("${security.saml.attribute-names.email}")
+	private String[] samlEmailAttributeNames;
+
+	@NotEmpty
+	@Value("${security.saml.attribute-names.group}")
+	private String[] samlGroupAttributeNames;
+	
+	@NotEmpty
+	@Value("${security.saml.attribute-names.details.office-state}")
+	private String[] samlOfficeStateAttributeNames;
 
 	@Override
 	public WaterAuthUser loadUserBySAML(SAMLCredential credential) throws UsernameNotFoundException {
-		return waterAuthUserFactory.createUserFromSamlAttributes(SAMLUtils.getAttributeValueMap(credential));
+		return createUserFromSamlAttributes(SAMLUtils.getAttributeValueMap(credential));
+	}
+
+	public WaterAuthUser createUserFromSamlAttributes(Map<String, List<String>> samlAttributeMap) {
+		return new WaterAuthUser(
+			SAMLUtils.getFirstMatchingAttributeFirstValue(samlAttributeMap, samlUsernameAttributeNames, true),
+			SAMLUtils.getFirstMatchingAttributeFirstValue(samlAttributeMap, samlEmailAttributeNames, true),
+			SAMLUtils.groupsToAuthoritiesList(samlAttributeMap, samlGroupAttributeNames),
+			createDetailsFromSamlAttributes(samlAttributeMap)
+		);
+	}
+
+	private WaterAuthUserDetails createDetailsFromSamlAttributes(Map<String, List<String>> samlAttributeMap) {
+		return new WaterAuthUserDetails(
+			SAMLUtils.getFirstMatchingAttributeFirstValue(samlAttributeMap, samlOfficeStateAttributeNames, false)
+		);
 	}
 }
